requirements.txt
PyYAML==6.0.3
Requests==2.32.5
